backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
bcrypt==4.0.1
pytest==7.4.2
PyJWT==2.8.0
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-Caching==2.1.0
Flask-Compress==1.13
Flask-Limiter==3.5.0
